requirements.txt
alabaster==0.7.12
asgiref==3.2.10
autopep8==1.5.7
Babel==2.8.0
beautifulsoup4==4.10.0
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
Django==3.1
django-filter==2.3.0
django-markdownx==3.0.1
djangorestframework==3.11.1
dnspython==1.16.0
docopt==0.6.2
docutils==0.16
fett==0.3.2
furo==2021.10.9
idna==2.10
imagesize==1.2.0
Jinja2==2.11.2
kiwisolver==1.3.1
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.4
mysqlclient==2.0.1
networkx==2.5
numpy==1.19.5
packaging==20.4
pandas==1.2.0
pathtools==0.1.2
Pillow==8.0.1
plotly==4.14.1
pycodestyle==2.7.0
Pygments==2.6.1
pymongo==3.11.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
retrying==1.3.3
six==1.15.0
snooty-lextudio==1.11.4.dev0
snowballstemmer==2.0.0
soupsieve==2.3.1
Sphinx==4.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.3.1
toml==0.10.2
typing-extensions==3.7.4.3
ujson==1.35
urllib3==1.25.10
uWSGI==2.0.19.1
watchdog==1.0.2
